# perf(ext/http): faster is_content_compressible