--- a/ext/http/Cargo.toml
+++ b/ext/http/Cargo.toml
@@ -30,6 +30,7 @@ fly-accept-encoding = "0.2.0-alpha.5"
 hyper = { version = "0.14.9", features = ["server", "stream", "http1", "http2", "runtime"] }
 mime = "0.3.16"
 percent-encoding = "2.1.0"
+phf = { version = "0.10", features = ["macros"] }
 ring = "0.16.20"
 serde = { version = "1.0.129", features = ["derive"] }
 tokio = { version = "1.17", features = ["full"] }