Cleanup + benches

Optimizing for the common case of "simple hits":
```
Before:
test compressible_complex_hit  ... bench:         113 ns/iter (+/- 3)
test compressible_complex_miss ... bench:         139 ns/iter (+/- 3)
test compressible_simple_hit   ... bench:          84 ns/iter (+/- 4)
test compressible_simple_miss  ... bench:          81 ns/iter (+/- 6)

After:
test compressible_complex_hit  ... bench:         126 ns/iter (+/- 1)
test compressible_complex_miss ... bench:         122 ns/iter (+/- 2)
test compressible_simple_hit   ... bench:          42 ns/iter (+/- 1)
test compressible_simple_miss  ... bench:         122 ns/iter (+/- 2)

After (with phf):
test compressible_complex_hit  ... bench:          76 ns/iter (+/- 1)
test compressible_complex_miss ... bench:          67 ns/iter (+/- 22)
test compressible_simple_hit   ... bench:          19 ns/iter (+/- 0)
test compressible_simple_miss  ... bench:          67 ns/iter (+/- 6)
```